Updates to patents search
