+++ main.py
@@ -7,7 +7,7 @@ from sympy import sequence
 
 from auto_db_pipeline.genbank.run_genbank_pipeline import run_genbank_pipeline
 from auto_db_pipeline.papers2ids import Papers
-from auto_db_pipeline.patents.patents_pipeline import get_seq
+from auto_db_pipeline.patents.patents_pipeline import get_seq_from_patents
 from parse_supp.get_supp_seqs import get_seqs_from_supp
 from auto_db_pipeline.get_additional_info.collate_results import collate_results
 from auto_db_pipeline.keywords_antigens import loadkeywords, load_known_antigens
@@ -23,7 +23,7 @@ def get_all_fucking_sequences():
 
 
   ''' Search for seqs from patents'''
-  patents = get_seq()
+  patents = get_seq_from_patents()
   ''' Search for seqs from papers '''
   # scrape paper text for pdb/genbank ids
   papers = Papers(selected_date="2022_03_08")
